Tests/LibWeb/Text/input/hit_testing/click-outside-of-box-with-lines-and-hidden-overflow.html

<script src="../include.js"></script>
<style>
    .overflow-hidden {
        overflow: hidden;
        border: 1px solid black;
        width: 200px;
        height: 200px;
    }

    #box {
        width: 400px;
        height: 400px;
    }
</style>
<div class="overflow-hidden">
    <div id="box">
        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse accumsan semper massa
        ut pharetra. Nunc vitae lorem at odio facilisis vehicula nec ut nibh. Duis ornare nibh orci,
        nec mollis est mollis ac. Aenean ultricies condimentum interdum. In eu elit vestibulum,
        sollicitudin lectus vel, congue nibh. Morbi vitae nunc in tortor ultrices aliquet sit amet a
        turpis. Phasellus ut dui sodales, dictum sem ut, efficitur tellus. Etiam eu orci et magna
        suscipit dignissim nec et sem. Aenean quam erat, euismod eu faucibus et, placerat ut dolor.
    </div>
</div>
<script type="text/javascript">
    test(() => {
        printElement(internals.hitTest(300, 300).node);
    });
</script>
